requirements.txt
affine==2.3.0
amply==0.1.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
boto3==1.17.88
botocore==1.20.88
certifi==2021.5.30
cffi==1.14.5
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
Fiona==1.8.20
geopandas==0.9.0
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json-tricks==3.15.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mock==4.0.3
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
PuLP==2.4
py==1.10.0
pycparser==2.20
pygdal==3.2.1.6
Pygments==2.9.0
pyparsing==2.4.7
pyproj==3.1.0
pyrsistent==0.17.3
pyshp==2.1.3
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
rasterio==1.2.4
rasterstats==0.15.0
richdem==0.3.4
s3transfer==0.4.2
scikit-criteria==0.2.10
scipy==1.6.3
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.17.2
six==1.16.0
snuggs==1.4.7
tabulate==0.8.9
terminado==0.10.0
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
